refactor(core): extract plugin_loader + utils to decepticon-core (Phase 1.B)
Per docs/superpowers/specs/2026-05-23-core-framework-sdk-split-design.md
§9.1 — second tranche of moves into the contract layer.

File moves (git mv preserves history):

  packages/decepticon/src/decepticon/plugin_loader.py
    -> packages/decepticon-core/src/decepticon_core/plugin_loader.py

  packages/decepticon/src/decepticon/core/config.py
    -> packages/decepticon-core/src/decepticon_core/utils/config.py

  packages/decepticon/src/decepticon/core/logging.py
    -> packages/decepticon-core/src/decepticon_core/utils/logging.py

All three were already langchain-free (stdlib + pydantic only).
``utils/config.py`` had one cross-package import to the framework
(``from decepticon.llm.models import ModelProfile``) — rewritten to
``from decepticon_core.types.llm import ModelProfile`` so the
contract layer no longer reaches back into the framework.

PluginBundle, SubAgentSpec, is_bundle_enabled, load_plugin_tools,
load_plugin_middleware, load_plugin_callbacks, load_plugin_skill_sources,
load_subagents_for_parent, load_plugin_agents — all reachable from
``decepticon_core.plugin_loader`` and (via shims) from
``decepticon.plugin_loader``. Identity preserved across paths.

Re-export shims at old paths keep framework call sites working
unchanged until Phase 2 rewrites them. Shims removed at 2.0.0 per
spec §7.3.

Tooling fix: workspace packages explicitly marked first-party in
``[tool.ruff.lint.isort]``. Without this, ruff's auto-detection
treated ``decepticon_core`` / ``decepticon_sdk`` as third-party,
triggering 41 I001 import-group violations on framework call sites
that legitimately mix the two namespaces. Marking them first-party
groups them together and clears the violations cleanly.

Phase 1.B gate:
  uv run ruff check .            ->  All checks passed!
  uv run ruff format --check .   ->  253 files already formatted
  uv run basedpyright --level error ->  0 errors, 0 warnings, 0 notes
  uv sync                        ->  119 packages, decepticon rebuilt
  identity check (legacy is core) ->  PluginBundle True, SubAgentSpec True

Remaining Phase 1 work:
  1.C: extract middleware_slots enum/constants (split — factories stay in framework)
  1.D: add new protocols (BackendProtocol, MiddlewareProtocol, ...)
  1.E: add new registries (RoleRegistry, SkillSourceRegistry, PluginRegistry.introspect)
  1.F: add new contributions + ruff banned-imports rule
